+++ NEWS.md
@@ -1,3 +1,15 @@
+# Rinvent 1.0.7
+
+## Documentation
+
+* Release 2023-03-11
+
+* readparquetR: partitioned parquet files can be read now
+* readparquetR: bug fix in add_part_names parameter
+* readparquetR: smart download in parquet format if partition is filled
+* write_to_blob: new function
+* date_mdy_ymd removed, date_to_yyyymmdd added
+
 # Rinvent 1.0.6
 
 ## Documentation